+++ src/response.js
@@ -25,9 +25,14 @@ export class ResponseWrapper {
         throw new Error('Not implemented');
     }
 
-    get body() {
+    get data() {
+        throw new Error('Not implemented');
+    }
+
+    get headers() {
         throw new Error('Not implemented');
     }
 }
+
 
 export default ResponseWrapper;