+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 # Changelog
 
+- [v1.0.0](#v100)
 - [v0.3.3](#v033)
 - [v0.3.1](#v031)
 - [v0.3.0](#v030)
@@ -9,6 +10,58 @@
 - [v0.1.0](#v010)
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
+
+### v1.0.0
+
+**v1.0.0 introduces breaking changes**
+
+ * Support react native (see #2)
+ * Error handling changes (see #3)
+ * No global configuration anymore
+ * Routers (see documentation)
+ * Misc
+  * Use babel 6 for building
+  * Reduce dependency count
+   * Use `lodash.template` instead of depending on full lodash
+
+#### Migrating to 1.0.0
+
+1. Default export changed:
+
+old: `import Resource from 'tg-resources';`
+new: `import { Resource } from 'tg-resources';`
+
+2. Global configuration was removed
+
+Instead of using `setConfig` and `getConfig` one must set options per Resource^1
+
+ [1] To keep things DRY use `Router` for defining your configuration. Note: It's also possible to extend `Resource` and can be a better
+ alternative in some cases.
+
+3. Configuration options have changed:
+
+ - `API_BASE` is now `apiRoot`
+ - `getExtraHeaders` is now `headers`
+ - `getCookies` is now `cookies`
+
+ - added:
+   - `mutateResponse`
+   - `prepareError`
+   - `parseErrors`
+   - `statusSuccess`
+   - `statusValidationError`
+   - `defaultHeaders`
+
+ - removed:
+   - `onSourceError`
+   - `ValidationErrorExtras`
+
+see the [Configuration](README.md#configuration) for more info
+
+4. Resource constructor changed:
+
+old: `new Resource(apiEndpoint, expectedStatus, mutateResponse, errorStatus)`
+new: `new Resource(apiEndpoint, options)`
 
 ### v0.3.3